Consolidate these client classes
